move a couple things out of evm bots
